refactor: split telemetry delta reader
